split flags handling is seperate files: flagsfilter.go, flagscapture.go and falgsoutput.go
