+++ cmd/tracee-ebpf/main.go
@@ -5,7 +5,6 @@ import (
 	"io"
 	"io/ioutil"
 	"log"
-	"net"
 	"os"
 	"os/exec"
 	"path"
@@ -54,8 +53,7 @@ func main() {
 			}
 
 			if checkCommandIsHelp(c.StringSlice("capture")) {
-				printCaptureHelp()
-				return nil
+				fmt.Print(captureHelp())
 			}
 			capture, err := prepareCapture(c.StringSlice("capture"))
 			if err != nil {
@@ -64,7 +62,7 @@ func main() {
 			cfg.Capture = &capture
 
 			if checkCommandIsHelp(c.StringSlice("trace")) {
-				printFilterHelp()
+				fmt.Print(filterHelp())
 				return nil
 			}
 			filter, err := prepareFilter(c.StringSlice("trace"))
@@ -77,7 +75,7 @@ func main() {
 				(cfg.Filter.NewContFilter.Enabled && cfg.Filter.NewContFilter.Value)
 
 			if checkCommandIsHelp(c.StringSlice("output")) {
-				printOutputHelp()
+				fmt.Print(outputHelp())
 				return nil
 			}
 			output, printerConfig, err := prepareOutput(c.StringSlice("output"))
@@ -286,7 +284,7 @@ func main() {
 				Name:    "list",
 				Aliases: []string{"l"},
 				Value:   false,
-				Usage:   "just list tracable events",
+				Usage:   "just list traceable events",
 			},
 			&cli.StringSliceFlag{
 				Name:    "trace",
@@ -304,7 +302,7 @@ func main() {
 				Name:    "output",
 				Aliases: []string{"o"},
 				Value:   cli.NewStringSlice("format:table"),
-				Usage:   "Control how and where output is printed. run '--output help' for more info.",
+				Usage:   "control how and where output is printed. run '--output help' for more info.",
 			},
 			&cli.IntFlag{
 				Name:    "perf-buffer-size",
@@ -351,525 +349,12 @@ func checkCommandIsHelp(s []string) bool {
 	return false
 }
 
-func printOutputHelp() {
-	outputHelp := `Control how and where output is printed.
-Possible options:
-
-[format:]table                                     output events in table format
-[format:]table-verbose                             output events in table format with extra fields per event
-[format:]json                                      output events in json format
-[format:]gob                                       output events in gob format
-[format:]gotemplate=/path/to/template              output events formatted using a given gotemplate file
-
-out-file:/path/to/file                             write the output to a specified file. create/trim the file if exists (default: stdout)
-err-file:/path/to/file                             write the errors to a specified file. create/trim the file if exists (default: stderr)
-
-none                                               ignore stream of events output, usually used with --capture
-
-option:{stack-addresses,detect-syscall,exec-env,relative-time,exec-hash,parse-arguments}
-                                                   augment output according to given options (default: none)
-  stack-addresses                                  include stack memory addresses for each event
-  detect-syscall                                   when tracing kernel functions which are not syscalls, detect and show the original syscall that called that function
-  exec-env                                         when tracing execve/execveat, show the environment variables that were used for execution
-  relative-time                                    use relative timestamp instead of wall timestamp for events
-  exec-hash                                        when tracing sched_process_exec, show the file hash(sha256) and ctime
-  parse-arguments                                  do not show raw machine-readable values for event arguments, instead parse into human readable strings
-
-Examples:
-  --output json                                            | output as json
-  --output gotemplate=/path/to/my.tmpl                     | output as the provided go template
-  --output out-file:/my/out --output err-file:/my/err      | output to /my/out and errors to /my/err
-  --output none                                            | ignore events output
-
-Use this flag multiple times to choose multiple output options
-`
-	fmt.Print(outputHelp)
-}
-
 type printerConfig struct {
 	kind    string
 	outPath string
 	outFile *os.File
 	errPath string
 	errFile *os.File
-}
-
-func prepareOutput(outputSlice []string) (tracee.OutputConfig, printerConfig, error) {
-	outcfg := tracee.OutputConfig{}
-	printcfg := printerConfig{}
-	printerKind := "table"
-	outPath := ""
-	errPath := ""
-	for _, o := range outputSlice {
-		outputParts := strings.SplitN(o, ":", 2)
-		numParts := len(outputParts)
-		if numParts == 1 && outputParts[0] != "none" {
-			outputParts = append(outputParts, outputParts[0])
-			outputParts[0] = "format"
-		}
-
-		switch outputParts[0] {
-		case "none":
-			printerKind = "ignore"
-		case "format":
-			printerKind = outputParts[1]
-			if printerKind != "table" &&
-				printerKind != "table-verbose" &&
-				printerKind != "json" &&
-				printerKind != "gob" &&
-				!strings.HasPrefix(printerKind, "gotemplate=") {
-				return outcfg, printcfg, fmt.Errorf("unrecognized output format: %s. Valid format values: 'table', 'table-verbose', 'json', 'gob' or 'gotemplate='. Use '--output help' for more info.", printerKind)
-			}
-		case "out-file":
-			outPath = outputParts[1]
-		case "err-file":
-			errPath = outputParts[1]
-		case "option":
-			switch outputParts[1] {
-			case "stack-addresses":
-				outcfg.StackAddresses = true
-			case "detect-syscall":
-				outcfg.DetectSyscall = true
-			case "exec-env":
-				outcfg.ExecEnv = true
-			case "relative-time":
-				outcfg.RelativeTime = true
-			case "exec-hash":
-				outcfg.ExecHash = true
-			case "parse-arguments":
-				outcfg.ParseArguments = true
-			default:
-				return outcfg, printcfg, fmt.Errorf("invalid output option: %s, use '--output help' for more info", outputParts[1])
-			}
-		default:
-			return outcfg, printcfg, fmt.Errorf("invalid output value: %s, use '--output help' for more info", outputParts[1])
-		}
-	}
-
-	if printerKind == "table" {
-		outcfg.ParseArguments = true
-	}
-
-	printcfg.kind = printerKind
-
-	if outPath == "" {
-		printcfg.outFile = os.Stdout
-	} else {
-		printcfg.outPath = outPath
-		fileInfo, err := os.Stat(outPath)
-		if err == nil {
-			if fileInfo.IsDir() {
-				return outcfg, printcfg, fmt.Errorf("cannot use a path of existing directory %s", outPath)
-			}
-		} else {
-			dir := filepath.Dir(outPath)
-			os.MkdirAll(dir, 0755)
-			printcfg.outFile, err = os.Create(outPath)
-			if err != nil {
-				return outcfg, printcfg, fmt.Errorf("failed to create output path: %v", err)
-			}
-		}
-	}
-
-	if errPath == "" {
-		printcfg.errFile = os.Stderr
-	} else {
-		printcfg.errPath = errPath
-		fileInfo, err := os.Stat(errPath)
-		if err == nil {
-			if fileInfo.IsDir() {
-				return outcfg, printcfg, fmt.Errorf("cannot use a path of existing directory %s", errPath)
-			}
-		} else {
-			dir := filepath.Dir(errPath)
-			os.MkdirAll(dir, 0755)
-			printcfg.errFile, err = os.Create(errPath)
-			if err != nil {
-				return outcfg, printcfg, fmt.Errorf("failed to create output path: %v", err)
-			}
-		}
-	}
-
-	return outcfg, printcfg, nil
-}
-
-func printCaptureHelp() {
-	captureHelp := `Capture artifacts that were written, executed or found to be suspicious.
-Captured artifacts will appear in the 'output-path' directory.
-Possible options:
-
-[artifact:]write[=/path/prefix*]   capture written files. A filter can be given to only capture file writes whose path starts with some prefix (up to 50 characters). Up to 3 filters can be given.
-[artifact:]exec                    capture executed files.
-[artifact:]module                  capture loaded kernel modules.
-[artifact:]mem                     capture memory regions that had write+execute (w+x) protection, and then changed to execute (x) only.
-[artifact:]net=interface           capture network traffic of the given interface. Only TCP/UDP protocols are currently supported.
-
-dir:/path/to/dir        path where tracee will save produced artifacts. the artifact will be saved into an 'out' subdirectory. (default: /tmp/tracee).
-profile                 creates a runtime profile of program executions and their metadata for forensics use.
-clear-dir               clear the captured artifacts output dir before starting (default: false).
-
-Examples:
-  --capture exec                                           | capture executed files into the default output directory
-  --capture exec --capture dir:/my/dir --capture clear-dir | delete /my/dir/out and then capture executed files into it
-  --capture write=/usr/bin/* --capture write=/etc/*        | capture files that were written into anywhere under /usr/bin/ or /etc/
-  --capture profile                                        | capture executed files and create a runtime profile in the output directory
-  --capture net=eth0                                       | capture network traffic of eth0
-  --capture exec --output none                             | capture executed files into the default output directory not printing the stream of events
-
-Use this flag multiple times to choose multiple capture options
-`
-	fmt.Print(captureHelp)
-}
-
-func prepareCapture(captureSlice []string) (tracee.CaptureConfig, error) {
-	capture := tracee.CaptureConfig{}
-
-	outDir := "/tmp/tracee"
-	clearDir := false
-
-	var filterFileWrite []string
-	for i := range captureSlice {
-		cap := captureSlice[i]
-		if strings.HasPrefix(cap, "artifact:write") ||
-			strings.HasPrefix(cap, "artifact:exec") ||
-			strings.HasPrefix(cap, "artifact:mem") ||
-			strings.HasPrefix(cap, "artifact:module") {
-			cap = strings.TrimPrefix(cap, "artifact:")
-		}
-		if cap == "write" {
-			capture.FileWrite = true
-		} else if strings.HasPrefix(cap, "write=") && strings.HasSuffix(cap, "*") {
-			capture.FileWrite = true
-			pathPrefix := strings.TrimSuffix(strings.TrimPrefix(cap, "write="), "*")
-			if len(pathPrefix) == 0 {
-				return tracee.CaptureConfig{}, fmt.Errorf("capture write filter cannot be empty")
-			}
-			filterFileWrite = append(filterFileWrite, pathPrefix)
-		} else if cap == "exec" {
-			capture.Exec = true
-		} else if cap == "module" {
-			capture.Module = true
-		} else if cap == "mem" {
-			capture.Mem = true
-		} else if strings.HasPrefix(cap, "net=") {
-			iface := strings.TrimPrefix(cap, "net=")
-			if _, err := net.InterfaceByName(iface); err != nil {
-				return tracee.CaptureConfig{}, fmt.Errorf("invalid network interface: %s", iface)
-			}
-			found := false
-			// Check if we already have this interface
-			for _, item := range capture.NetIfaces {
-				if iface == item {
-					found = true
-					break
-				}
-			}
-			if !found {
-				capture.NetIfaces = append(capture.NetIfaces, iface)
-			}
-		} else if cap == "clear-dir" {
-			clearDir = true
-		} else if strings.HasPrefix(cap, "dir:") {
-			outDir = strings.TrimPrefix(cap, "dir:")
-			if len(outDir) == 0 {
-				return tracee.CaptureConfig{}, fmt.Errorf("capture output dir cannot be empty")
-			}
-		} else if cap == "profile" {
-			capture.Exec = true
-			capture.Profile = true
-		} else {
-			return tracee.CaptureConfig{}, fmt.Errorf("invalid capture option specified, use '--capture help' for more info")
-		}
-	}
-	capture.FilterFileWrite = filterFileWrite
-
-	capture.OutputPath = filepath.Join(outDir, "out")
-	if clearDir {
-		os.RemoveAll(capture.OutputPath)
-	}
-
-	return capture, nil
-}
-
-func printFilterHelp() {
-	filterHelp := `Select which events to trace by defining trace expressions that operate on events or process metadata.
-Only events that match all trace expressions will be traced (trace flags are ANDed).
-The following types of expressions are supported:
-
-Numerical expressions which compare numbers and allow the following operators: '=', '!=', '<', '>'.
-Available numerical expressions: uid, pid, mntns, pidns.
-
-String expressions which compares text and allow the following operators: '=', '!='.
-Available string expressions: event, set, uts, comm.
-
-Boolean expressions that check if a boolean is true and allow the following operator: '!'.
-Available boolean expressions: container.
-
-Event arguments can be accessed using 'event_name.event_arg' and provide a way to filter an event by its arguments.
-Event arguments allow the following operators: '=', '!='.
-Strings can be compared as a prefix if ending with '*'.
-
-Event return value can be accessed using 'event_name.retval' and provide a way to filter an event by its return value.
-Event return value expression has the same syntax as a numerical expression.
-
-Non-boolean expressions can compare a field to multiple values separated by ','.
-Multiple values are ORed if used with equals operator '=', but are ANDed if used with any other operator.
-
-The field 'container' and 'pid' also support the special value 'new' which selects new containers or pids, respectively.
-
-The field 'set' selects a set of events to trace according to predefined sets, which can be listed by using the 'list' flag.
-
-The special 'follow' expression declares that not only processes that match the criteria will be traced, but also their descendants.
-
-Examples:
-  --trace pid=new                                              | only trace events from new processes
-  --trace pid=510,1709                                         | only trace events from pid 510 or pid 1709
-  --trace p=510 --trace p=1709                                 | only trace events from pid 510 or pid 1709 (same as above)
-  --trace container=new                                        | only trace events from newly created containers
-  --trace container                                            | only trace events from containers
-  --trace c                                                    | only trace events from containers (same as above)
-  --trace '!container'                                         | only trace events from the host
-  --trace uid=0                                                | only trace events from uid 0
-  --trace mntns=4026531840                                     | only trace events from mntns id 4026531840
-  --trace pidns!=4026531836                                    | only trace events from pidns id not equal to 4026531840
-  --trace tree=476165                                          | only trace events that descend from the process with pid 476165
-  --trace tree!=5023                                           | only trace events if they do not descend from the process with pid 5023
-  --trace tree=3213,5200 --trace tree!=3215                    | only trace events if they descend from 3213 or 5200, but not 3215
-  --trace 'uid>0'                                              | only trace events from uids greater than 0
-  --trace 'pid>0' --trace 'pid<1000'                           | only trace events from pids between 0 and 1000
-  --trace 'u>0' --trace u!=1000                                | only trace events from uids greater than 0 but not 1000
-  --trace event=execve,open                                    | only trace execve and open events
-  --trace event=open*                                          | only trace events prefixed by "open"
-  --trace event!=open*,dup*                                    | don't trace events prefixed by "open" or "dup"
-  --trace set=fs                                               | trace all file-system related events
-  --trace s=fs --trace e!=open,openat                          | trace all file-system related events, but not open(at)
-  --trace uts!=ab356bc4dd554                                   | don't trace events from uts name ab356bc4dd554
-  --trace comm=ls                                              | only trace events from ls command
-  --trace close.fd=5                                           | only trace 'close' events that have 'fd' equals 5
-  --trace openat.pathname=/tmp*                                | only trace 'openat' events that have 'pathname' prefixed by "/tmp"
-  --trace openat.pathname!=/tmp/1,/bin/ls                      | don't trace 'openat' events that have 'pathname' equals /tmp/1 or /bin/ls
-  --trace comm=bash --trace follow                             | trace all events that originated from bash or from one of the processes spawned by bash
-
-
-Note: some of the above operators have special meanings in different shells.
-To 'escape' those operators, please use single quotes, e.g.: 'uid>0'
-`
-	fmt.Print(filterHelp)
-}
-
-func prepareFilter(filters []string) (tracee.Filter, error) {
-	filter := tracee.Filter{
-		UIDFilter: &tracee.UintFilter{
-			Equal:    []uint64{},
-			NotEqual: []uint64{},
-			Less:     tracee.LessNotSetUint,
-			Greater:  tracee.GreaterNotSetUint,
-			Is32Bit:  true,
-		},
-		PIDFilter: &tracee.UintFilter{
-			Equal:    []uint64{},
-			NotEqual: []uint64{},
-			Less:     tracee.LessNotSetUint,
-			Greater:  tracee.GreaterNotSetUint,
-			Is32Bit:  true,
-		},
-		NewPidFilter: &tracee.BoolFilter{},
-		MntNSFilter: &tracee.UintFilter{
-			Equal:    []uint64{},
-			NotEqual: []uint64{},
-			Less:     tracee.LessNotSetUint,
-			Greater:  tracee.GreaterNotSetUint,
-		},
-		PidNSFilter: &tracee.UintFilter{
-			Equal:    []uint64{},
-			NotEqual: []uint64{},
-			Less:     tracee.LessNotSetUint,
-			Greater:  tracee.GreaterNotSetUint,
-		},
-		UTSFilter: &tracee.StringFilter{
-			Equal:    []string{},
-			NotEqual: []string{},
-		},
-		CommFilter: &tracee.StringFilter{
-			Equal:    []string{},
-			NotEqual: []string{},
-		},
-		ContFilter:    &tracee.BoolFilter{},
-		NewContFilter: &tracee.BoolFilter{},
-		RetFilter: &tracee.RetFilter{
-			Filters: make(map[int32]tracee.IntFilter),
-		},
-		ArgFilter: &tracee.ArgFilter{
-			Filters: make(map[int32]map[string]tracee.ArgFilterVal),
-		},
-		ProcessTreeFilter: &tracee.ProcessTreeFilter{
-			PIDs: make(map[uint32]bool),
-		},
-		EventsToTrace: []int32{},
-	}
-
-	eventFilter := &tracee.StringFilter{Equal: []string{}, NotEqual: []string{}}
-	setFilter := &tracee.StringFilter{Equal: []string{}, NotEqual: []string{}}
-
-	eventsNameToID := make(map[string]int32, len(tracee.EventsIDToEvent))
-	for _, event := range tracee.EventsIDToEvent {
-		eventsNameToID[event.Name] = event.ID
-	}
-
-	for _, f := range filters {
-		filterName := f
-		operatorAndValues := ""
-		operatorIndex := strings.IndexAny(f, "=!<>")
-		if operatorIndex > 0 {
-			filterName = f[0:operatorIndex]
-			operatorAndValues = f[operatorIndex:]
-		}
-
-		if strings.Contains(f, ".retval") {
-			err := filter.RetFilter.Parse(filterName, operatorAndValues, eventsNameToID)
-			if err != nil {
-				return tracee.Filter{}, err
-			}
-			continue
-		}
-
-		if strings.Contains(f, ".") {
-			err := filter.ArgFilter.Parse(filterName, operatorAndValues, eventsNameToID)
-			if err != nil {
-				return tracee.Filter{}, err
-			}
-			continue
-		}
-
-		// The filters which are more common (container, event, pid, set, uid) can be given using a prefix of them.
-		// Other filters should be given using their full name.
-		// To avoid collisions between filters that share the same prefix, put the filters which should have an exact match first!
-		if filterName == "comm" {
-			err := filter.CommFilter.Parse(operatorAndValues)
-			if err != nil {
-				return tracee.Filter{}, err
-			}
-			continue
-		}
-
-		if strings.HasPrefix("container", f) || (strings.HasPrefix("!container", f) && len(f) > 1) {
-			filter.NewPidFilter.Enabled = true
-			filter.NewPidFilter.Value = true
-			err := filter.ContFilter.Parse(f)
-			if err != nil {
-				return tracee.Filter{}, err
-			}
-			continue
-		}
-
-		if strings.HasPrefix("container", filterName) {
-			if operatorAndValues == "=new" {
-				filter.NewPidFilter.Enabled = true
-				filter.NewPidFilter.Value = true
-				filter.NewContFilter.Enabled = true
-				filter.NewContFilter.Value = true
-				continue
-			}
-			if operatorAndValues == "!=new" {
-				filter.ContFilter.Enabled = true
-				filter.ContFilter.Value = true
-				filter.NewPidFilter.Enabled = true
-				filter.NewPidFilter.Value = true
-				filter.NewContFilter.Enabled = true
-				filter.NewContFilter.Value = false
-				continue
-			}
-		}
-
-		if strings.HasPrefix("event", filterName) {
-			err := eventFilter.Parse(operatorAndValues)
-			if err != nil {
-				return tracee.Filter{}, err
-			}
-			continue
-		}
-
-		if filterName == "mntns" {
-			err := filter.MntNSFilter.Parse(operatorAndValues)
-			if err != nil {
-				return tracee.Filter{}, err
-			}
-			continue
-		}
-
-		if filterName == "pidns" {
-			err := filter.PidNSFilter.Parse(operatorAndValues)
-			if err != nil {
-				return tracee.Filter{}, err
-			}
-			continue
-		}
-
-		if filterName == "tree" {
-			err := filter.ProcessTreeFilter.Parse(operatorAndValues)
-			if err != nil {
-				return tracee.Filter{}, err
-			}
-			continue
-		}
-
-		if strings.HasPrefix("pid", filterName) {
-			if operatorAndValues == "=new" {
-				filter.NewPidFilter.Enabled = true
-				filter.NewPidFilter.Value = true
-				continue
-			}
-			if operatorAndValues == "!=new" {
-				filter.NewPidFilter.Enabled = true
-				filter.NewPidFilter.Value = false
-				continue
-			}
-			err := filter.PIDFilter.Parse(operatorAndValues)
-			if err != nil {
-				return tracee.Filter{}, err
-			}
-			continue
-		}
-
-		if strings.HasPrefix("set", filterName) {
-			err := setFilter.Parse(operatorAndValues)
-			if err != nil {
-				return tracee.Filter{}, err
-			}
-			continue
-		}
-
-		if filterName == "uts" {
-			err := filter.UTSFilter.Parse(operatorAndValues)
-			if err != nil {
-				return tracee.Filter{}, err
-			}
-			continue
-		}
-
-		if strings.HasPrefix("uid", filterName) {
-			err := filter.UIDFilter.Parse(operatorAndValues)
-			if err != nil {
-				return tracee.Filter{}, err
-			}
-			continue
-		}
-
-		if strings.HasPrefix("follow", f) {
-			filter.Follow = true
-			continue
-		}
-
-		return tracee.Filter{}, fmt.Errorf("invalid filter option specified, use '--trace help' for more info")
-	}
-
-	var err error
-	filter.EventsToTrace, err = prepareEventsToTrace(eventFilter, setFilter, eventsNameToID)
-	if err != nil {
-		return tracee.Filter{}, err
-	}
-
-	return filter, nil
 }
 
 func prepareEventsToTrace(eventFilter *tracee.StringFilter, setFilter *tracee.StringFilter, eventsNameToID map[string]int32) ([]int32, error) {
